MAJOR cleaning and restructuring
I finally decided to clean "some" stuff, aka a damn lot. I don't really like that I've done this all together, but I was already too deep into it when I noticed and it's too much to separate now...
A summary:
- completed some TODOs
- got rid of small custom messages in favor of a unique one
- changed some datatypes, especially in functions' arguments
- better input handling in Datapad
- migrated some constants to ROS2 parameters
- defined some custom formatter to clean logs
- double-checked 'const' declarations
- migrated from 'float' to 'double'
- tried to improve statements and code usage (e.g. vector.empty() instead of vector.size>0)
- substituted manual time-based function calls with timers, since this is ROS2
- deleted everything that was unused

+++ src/pelican/src/TacMapModule/external_comms.cpp
@@ -59,7 +59,7 @@ rclcpp::SubscriptionOptions TacMapModule::gatherReentrantOptions() const {
 }
 
 /************* To make other modules carry on an action ************/
-void TacMapModule::signalHeightCompensation(float odom_height) const {
+void TacMapModule::signalHeightCompensation(double odom_height) const {
     if (!this->node_) {
         throw MissingExternModule();
     }
